Fixing bugs
Had it read the data in a similar manner to the way the real thing does it rather than pre-dividing it into 100-length chunks. Also had it account for offset in when it performs a calculation and starts a new buffer. Also changed code to look more like the Teensy code and remove NumPy optimizations (not all of them).